new file for base predictions
